Add files via upload
This a pong game in c run on the nios 2 processor for the DE10 nano board. Quatros has been use to compile and program the device.
The cpu is connected to and hdmi module which is responsible of diplaying the game on the screen. 
A custom made shield for the nano has been used to display the score on 4 7-segment diplays.
Furthermore, to connect the controllers and the speaker. The controller works fine but the audio module is not included.